requirements.txt
torch==1.2
numpy
pypinyin
OpenCC
Pillow
tqdm
